:sparkles: move runnable modules to exported functions
the problem with having a "runnable module" (to be able to run a module without importing a *function* from it) is that it is impossible to mock test. If you run functions inside modules, frameworks like Jest or Bun.test have the possiblity to replace entire imports. This resolves in the ability to mock responses of APIs and DBs anywhere